Introduce lpmd_misc.c
Introduce lpmd_misc.c for code that handling misc power features
including ITMT and EPP/EPB.
