+++ Makefile.am
@@ -59,6 +59,7 @@ intel_lpmd_SOURCES = \
 	src/lpmd_socket.c \
 	src/lpmd_util.c \
 	src/lpmd_wlt.c \
+	src/lpmd_misc.c \
 	src/lpmd_state_machine.c \
 	src/wlt_proxy/wlt_proxy.c \
 	src/wlt_proxy/spike_mgmt.c \
