+++ src/lpmd_cpu.c
@@ -466,235 +466,6 @@ char *get_irqbalance_str(enum cpumask_idx idx)
 	return get_cpus_str_reverse(idx);
 }
 
-/* Handling EPP */
-
-#define MAX_EPP_STRING_LENGTH	32
-struct cpu_info {
-	char epp_str[MAX_EPP_STRING_LENGTH];
-	int epp;
-	int epb;
-};
-static struct cpu_info *saved_cpu_info;
-
-static int lp_mode_epp = SETTING_IGNORE;
-
-int get_lpm_epp(void)
-{
-	return lp_mode_epp;
-}
-
-void set_lpm_epp(int val)
-{
-	lp_mode_epp = val;
-}
-
-static int lp_mode_epb = SETTING_IGNORE;
-
-int get_lpm_epb(void)
-{
-	return lp_mode_epb;
-}
-
-void set_lpm_epb(int val)
-{
-	lp_mode_epb = val;
-}
-
-int get_epp(char *path, int *val, char *str, int size)
-{
-	FILE *filep;
-	int epp;
-	int ret;
-
-	filep = fopen (path, "r");
-	if (!filep)
-		return 1;
-
-	ret = fscanf (filep, "%d", &epp);
-	if (ret == 1) {
-		*val = epp;
-		ret = 0;
-		goto end;
-	}
-
-	ret = fread (str, 1, size, filep);
-	if (ret <= 0)
-		ret = 1;
-	else {
-		if (ret >= size)
-			ret = size - 1;
-		str[ret - 1] = '\0';
-		ret = 0;
-	}
-end:
-	fclose (filep);
-	return ret;
-}
-
-int set_epp(char *path, int val, char *str)
-{
-	FILE *filep;
-	int ret;
-
-	filep = fopen (path, "r+");
-	if (!filep)
-		return 1;
-
-	if (val >= 0)
-		ret = fprintf (filep, "%d", val);
-	else if (str && str[0] != '\0')
-		ret = fprintf (filep, "%s", str);
-	else {
-		fclose (filep);
-		return 1;
-	}
-
-	fclose (filep);
-
-	if (ret <= 0) {
-		if (val >= 0)
-			lpmd_log_error ("Write \"%d\" to %s failed, ret %d\n", val, path, ret);
-		else
-			lpmd_log_error ("Write \"%s\" to %s failed, ret %d\n", str, path, ret);
-	}
-	return !(ret > 0);
-}
-
-static char *get_ppd_default_epp(void)
-{
-	int ppd_mode = get_ppd_mode();
-
-	if (ppd_mode == PPD_INVALID)
-		return NULL;
-
-	if (ppd_mode == PPD_PERFORMANCE)
-		return "performance";
-
-	if (ppd_mode == PPD_POWERSAVER)
-		return "power";
-
-	if (is_on_battery())
-		return "balance_power";
-
-	return "balance_performance";
-}
-
-int get_epp_epb(int *epp, char *epp_str, int size, int *epb)
-{
-	int c;
-	char path[MAX_STR_LENGTH];
-
-	for (c = 0; c < max_online_cpu; c++) {
-		if (!is_cpu_online (c))
-			continue;
-
-		*epp = -1;
-		epp_str[0] = '\0';
-		snprintf (path, sizeof(path), "/sys/devices/system/cpu/cpu%d/cpufreq/energy_performance_preference", c);
-		get_epp (path, epp, epp_str, size);
-
-		snprintf(path, MAX_STR_LENGTH, "/sys/devices/system/cpu/cpu%d/power/energy_perf_bias", c);
-		lpmd_read_int(path, epb, -1);
-		return 0;
-	}
-	return 1;
-}
-
-int init_epp_epb(void)
-{
-	int max_cpus = get_max_cpus ();
-	int c;
-	int ret;
-	char path[MAX_STR_LENGTH];
-
-	saved_cpu_info = calloc (max_cpus, sizeof(struct cpu_info));
-
-	for (c = 0; c < max_cpus; c++) {
-		saved_cpu_info[c].epp_str[0] = '\0';
-		saved_cpu_info[c].epp = -1;
-
-		if (!is_cpu_online (c))
-			continue;
-
-		snprintf (path, sizeof(path), "/sys/devices/system/cpu/cpu%d/cpufreq/energy_performance_preference", c);
-		ret = get_epp (path, &saved_cpu_info[c].epp, saved_cpu_info[c].epp_str, MAX_EPP_STRING_LENGTH);
-		if (!ret) {
-			if (saved_cpu_info[c].epp != -1)
-				lpmd_log_debug ("CPU%d EPP: 0x%x\n", c, saved_cpu_info[c].epp);
-			else
-				lpmd_log_debug ("CPU%d EPP: %s\n", c, saved_cpu_info[c].epp_str);
-		}
-
-		snprintf(path, MAX_STR_LENGTH, "/sys/devices/system/cpu/cpu%d/power/energy_perf_bias", c);
-		ret = lpmd_read_int(path, &saved_cpu_info[c].epb, -1);
-		if (ret) {
-			saved_cpu_info[c].epb = -1;
-			continue;
-		}
-		lpmd_log_debug ("CPU%d EPB: 0x%x\n", c, saved_cpu_info[c].epb);
-	}
-	return 0;
-}
-
-int process_epp_epb(void)
-{
-	int max_cpus = get_max_cpus ();
-	int c;
-	int ret;
-	char path[MAX_STR_LENGTH];
-
-	if (lp_mode_epp == SETTING_IGNORE)
-		lpmd_log_info ("Ignore EPP\n");
-	if (lp_mode_epb == SETTING_IGNORE)
-		lpmd_log_info ("Ignore EPB\n");
-	if (lp_mode_epp == SETTING_IGNORE && lp_mode_epb == SETTING_IGNORE)
-		return 0;
-
-	for (c = 0; c < max_cpus; c++) {
-		int val;
-		char *str = NULL;
-
-		if (!is_cpu_online (c))
-			continue;
-
-		if (lp_mode_epp != SETTING_IGNORE) {
-			if (lp_mode_epp == SETTING_RESTORE) {
-				val = -1;
-				str = get_ppd_default_epp();
-				if (!str) {
-					/* Fallback to cached EPP */
-					val = saved_cpu_info[c].epp;
-					str = saved_cpu_info[c].epp_str;
-				}
-			} else {
-				val = lp_mode_epp;
-			}
-
-			snprintf (path, sizeof(path), "/sys/devices/system/cpu/cpu%d/cpufreq/energy_performance_preference", c);
-			ret = set_epp (path, val, str);
-			if (!ret) {
-				if (val != -1)
-					lpmd_log_debug ("Set CPU%d EPP to 0x%x\n", c, val);
-				else
-					lpmd_log_debug ("Set CPU%d EPP to %s\n", c, saved_cpu_info[c].epp_str);
-			}
-		}
-
-		if (lp_mode_epb != SETTING_IGNORE) {
-			if (lp_mode_epb == SETTING_RESTORE)
-				val = saved_cpu_info[c].epb;
-			else
-				val = lp_mode_epb;
-
-			snprintf (path, MAX_STR_LENGTH, "/sys/devices/system/cpu/cpu%d/power/energy_perf_bias", c);
-			ret = lpmd_write_int(path, val, -1);
-			if (!ret)
-				lpmd_log_debug ("Set CPU%d EPB to 0x%x\n", c, val);
-		}
-	}
-	return 0;
-}
-
 static int uevent_fd = -1;
 
 int uevent_init(void)
@@ -1661,7 +1432,6 @@ int init_cpu(char *cmd_cpus, enum lpm_cpu_process_mode mode, int epp)
 	if (ret)
 		return ret;
 
-	init_epp_epb ();
 	return 0;
 }
 
@@ -1671,8 +1441,6 @@ int process_cpus(int enter, enum lpm_cpu_process_mode mode)
 
 	if (enter != 1 && enter != 0)
 		return LPMD_ERROR;
-
-	process_epp_epb ();
 
 	if (lpm_cpus_cur == CPUMASK_MAX) {
 		lpmd_log_info ("Ignore Task migration\n");
